Fix total score data type

+++ osu.ElasticIndexer/SoloScore.cs
@@ -84,7 +84,7 @@ namespace osu.ElasticIndexer
 
         [Computed]
         [Number(NumberType.Integer)]
-        public int total_score => scoreData.LegacyTotalScore ?? scoreData.TotalScore;
+        public int total_score => scoreData.LegacyTotalScore ?? (int)scoreData.TotalScore; // scoreData.TotalScore should never exceed int.MaxValue at the point of storage.
 
         [Computed]
         [Number(NumberType.Float)]
